EY-1095 Flytt restmodules til applicationbuilder og pseudo kode for service
